FSCORE-245 fix windows build in switch_generate_sln_silence
git-svn-id: http://svn.freeswitch.org/svn/freeswitch/trunk@10625 d0543943-73ff-0310-b7d9-9358b9ac24b2

+++ src/switch_resample.c
@@ -209,11 +209,11 @@ SWITCH_DECLARE(void) switch_generate_sln_silence(int16_t *data, uint32_t samples
 
 	assert(divisor);
 
-	rnd2 = (int16_t) (intptr_t) &data + switch_timestamp(NULL);
+	rnd2 = (int16_t) (intptr_t) &data + (int16_t)switch_timestamp(NULL);
 
 	for (i = 0; i < samples; i++, sum_rnd = 0) {
 		for (x = 0; x < 10; x++) {
-			rnd += ((x + i) * rnd2);
+			rnd += (int16_t)((x + i) * rnd2);
 			sum_rnd += rnd;
 		}
 		switch_normalize_to_16bit(sum_rnd);
